requirements.txt
fastapi
pydantic
pydantic[email]
uvicorn
sqlalchemy
pyjwt
python-decouple
asyncopg
